# sdk/openapi/external-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cross-border Agent External Integration API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/external/oauth/token": {
      "post": {
        "tags": [
          "external-v1"
        ],
        "summary": "Token",
        "operationId": "token_api_v1_external_oauth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_token_api_v1_external_oauth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenView"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "403": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "404": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "409": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "413": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "429": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "503": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "504": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/external/projects": {
      "post": {
        "tags": [
          "external-v1"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_external_projects_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeView"
                }
              }
            }
          },
          "422": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "403": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "404": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "409": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "413": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "429": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "503": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "504": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "IntegrationBearer": []
          }
        ]
      }
    },
...
    "/api/v1/external/projects/{project_id}/intake/confirm": {
      "post": {
        "tags": [
          "external-v1"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_external_projects__project_id__intake_confirm_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmProjectIntake"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeView"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "401": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "403": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "404": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "409": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "413": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "429": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "503": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          },
          "504": {
            "description": "Stable sanitized gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayError"
                }
              }
            }
          }
        },
        "security": [
          {
            "IntegrationBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_token_api_v1_external_oauth_token_post": {
        "properties": {
          "grant_type": {
            "type": "string",
            "title": "Grant Type"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "writeOnly": true,
            "format": "password"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "grant_type",
          "client_id",
          "client_secret"
        ],
        "title": "Body_token_api_v1_external_oauth_token_post",
        "additionalProperties": false
      },
...
      "ExternalProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "facts": {
            "$ref": "#/components/schemas/IntakeFacts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "facts"
        ],
        "title": "ExternalProjectCreate"
      },
...
      "IntakeView": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "project_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Version Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "CONFIRMED",
              "SUPERSEDED"
            ],
            "title": "Status"
          },
          "intake": {
            "$ref": "#/components/schemas/ProjectIntakeContext"
          },
          "analysis_snapshot_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Snapshot Id"
          },
          "workflow_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Run Id"
          },
          "retrieval_policy_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Policy Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "project_version_id",
          "version",
          "status",
          "intake"
        ],
        "title": "IntakeView"
      },
...
      "TokenView": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "scope"
        ],
        "title": "TokenView"
      },
...
      "GatewayError": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "error_code",
          "message",
          "trace_id"
        ],
        "title": "GatewayError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "IntegrationBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived OAuth token or revocable service credential."
      },
      "ClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/api/v1/external/oauth/token",
            "scopes": {
              "project:create": "project:create",
              "project:read": "project:read",
              "intake:write": "intake:write",
              "document:upload": "document:upload",
              "compliance:analyze": "compliance:analyze",
              "compliance:read": "compliance:read",
              "model:select": "model:select",
              "webhook:manage": "webhook:manage"
            }
          }
        }
      }
    }
  },
...
}